updated method deletePrescription() in PrescriptionService

+++ src/main/java/com/example/PrescriptionCRUD/services/PrescriptionService.java
@@ -17,6 +17,7 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -77,7 +78,12 @@ public class PrescriptionService {
 
     @Transactional
     public void deletePrescription(Long prescriptionId){
-        prescriptionRepository.deleteById(prescriptionId);
+        Optional<Prescription> deleteById = prescriptionRepository.findById(prescriptionId);
+        if (deleteById.isPresent()){
+            prescriptionRepository.deleteById(prescriptionId);
+        } else {
+            throw new IllegalStateException("Prescription with id:" + prescriptionId + " does not exist in database!");
+        }
     }
 
 }
